requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dill==0.3.8
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.1
ibm-cloud-sdk-core==3.20.1
ibm-platform-services==0.53.7
idna==3.7
mpmath==1.3.0
numpy==1.26.4
pbr==6.0.0
psutil==5.9.8
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
pyspnego==0.10.2
python-dateutil==2.9.0.post0
qiskit==1.1.0
qiskit-aer==0.14.2
qiskit-ibm-runtime==0.23.0
requests==2.32.2
requests-ntlm==1.2.0
rustworkx==0.14.2
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
sspilib==0.1.0
stevedore==5.2.0
symengine==0.11.0
sympy==1.12.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
